refactor: use type state for ptrace tracer
