+++ src/tui/app.rs
@@ -41,7 +41,7 @@ use crate::{
   event::{Event, ProcessStateUpdate, ProcessStateUpdateEvent, TracerEventDetails, TracerMessage},
   printer::PrinterArgs,
   proc::BaselineInfo,
-  ptrace::Tracer,
+  ptrace::RunningTracer,
   pty::{PtySize, UnixMasterPty},
   tui::{error_popup::InfoPopupState, query::QueryKind},
 };
@@ -78,7 +78,7 @@ pub struct App {
   pub should_handle_internal_resize: bool,
   pub popup: Vec<ActivePopup>,
   pub active_experiments: Vec<&'static str>,
-  tracer: Option<Arc<Tracer>>,
+  tracer: Option<RunningTracer>,
   query_builder: Option<QueryBuilder>,
   breakpoint_manager: Option<BreakPointManagerState>,
   hit_manager_state: Option<HitManagerState>,
@@ -86,7 +86,7 @@ pub struct App {
 }
 
 pub struct PTracer {
-  pub tracer: Arc<Tracer>,
+  pub tracer: RunningTracer,
   pub debugger_args: DebuggerArgs,
 }
 
@@ -211,9 +211,10 @@ impl App {
               }
               // Cancel hit manager
               if let Some(h) = self.hit_manager_state.as_mut()
-                && h.visible {
-                  h.hide();
-                }
+                && h.visible
+              {
+                h.hide();
+              }
               // action_tx.send(Action::Render)?;
             } else {
               trace!("TUI: Active pane: {}", self.active_pane);
@@ -250,12 +251,13 @@ impl App {
 
                 // Handle hit manager
                 if let Some(h) = self.hit_manager_state.as_mut()
-                  && h.visible {
-                    if let Some(action) = h.handle_key_event(ke) {
-                      action_tx.send(action);
-                    }
-                    continue;
+                  && h.visible
+                {
+                  if let Some(action) = h.handle_key_event(ke) {
+                    action_tx.send(action);
                   }
+                  continue;
+                }
 
                 // Handle breakpoint manager
                 if let Some(breakpoint_manager) = self.breakpoint_manager.as_mut() {
